Fixing delete routes

+++ app/crud/resumes.py
@@ -1,4 +1,5 @@
 import os
+import shutil
 from typing import Optional
 
 from sqlalchemy.orm import Session
@@ -45,6 +46,10 @@ def get_resumes_by_tag_id(db: Session, tag_id: int, user_id: int) -> Optional[li
     return None
 
 def delete_all_resumes(db: Session) -> bool:
+    if os.path.exists('app/data/'):
+        shutil.rmtree('app/data/')
+
+    os.makedirs('app/data') 
     db.query(models.Resume).delete()
     db.commit()
     return True
@@ -54,7 +59,7 @@ def update_resume(db: Session, resume: schemas.ResumeUpdate, user: schemas.User)
 
     for field, value in vars(resume).items():
         setattr(db_resume, field, value) if value else None
-        
+
     db.add(db_resume)
     db.commit()
     db.refresh(db_resume)
@@ -62,6 +67,9 @@ def update_resume(db: Session, resume: schemas.ResumeUpdate, user: schemas.User)
 
 def delete_resume(db: Session, resume: models.Resume):
     os.remove(resume.filename)
+    if os.path.exists(resume.filename):
+        return False
+
     db.delete(resume)
     db.commit()
-    return resume.object_id
+    return resume.id
